Footers Fixed!
The footers have been fixed.  The one on the contact page had a duplicate of the footer in the wrapper div, so it extended until the end so i removed that. On the events page, I missed a div so the closing div tag went into a div inside of the wrapper div, so the wrapper div didn't close. They should both work now.